update tests, add relay blockPublishing
