+++ Example/ExampleApp.xcodeproj/IntegrationTests.xctestplan
@@ -54,6 +54,7 @@
     {
       "parallelizable" : true,
       "skippedTests" : [
+        "AuthTests",
         "AuthTests\/testEIP1271RespondSuccess()",
         "ChatTests",
         "ENSResolverTests",
